Renaming of dependency crates
A mechanism for "renaming" crate imports is needed for when a dependant
crate is renamed.
See "Renaming dependencies in Cargo.toml"("Specifying Dependencies")
in the Cargo Book.

It is very simplistic for now: just put a string before the reference
to the dependency
Save up these renamings in an array in a pass before handling
the dependencies.