Propagate link lib search directives properly
Link lib search directories info, etc is added to the depinfo files.
It will then be picked up from the direct dependencies.
However, sometimes this info is needed further the down the tree of
dependants. E.g. to do the final link.
If the dependency that needs a lib is not a direct dependency, the build
will break.
Fix by appending the depinfo files of dependencies also, when creating
a depinfo file.
The downside is that the link search directive can be duplicated many
times.
However, apart from being ugly, it doesn't seem to have any negative
effect.